refactor(bayn): modularize startup workflow
